Añadida funcionalidad de subida de ficheros de imagen, y su eliminación correcta.

+++ app/Http/Controllers/ImagesController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Images;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\File;
 
 class ImagesController extends Controller
 {
@@ -102,9 +103,11 @@ class ImagesController extends Controller
     {
         $image = Images::find($id);
 
-        $image->delete();
-
-        return redirect()->route('images.index')
-            ->with('success', 'Imagen borrada');
+        if ($image) {
+            $imgPath = public_path('productImages/' . $image->imgURL);
+            if (File::exists($imgPath)) File::delete($imgPath);
+            $image->delete();
+            return redirect()->route('images.index')->with('success', 'Imagen borrada');
+        }
     }
 }
